+++ test/sampler/test_sampler.py
@@ -28,7 +28,6 @@ from netket import config
 from netket.hilbert import DiscreteHilbert
 from netket.experimental.hilbert import Particle
 from netket.utils import array_in, mpi
-from netket.jax.sharding import device_count_per_rank
 
 
 from .. import common
@@ -611,7 +610,7 @@ def test_exact_sampler(sampler):
         assert sampler.n_chains_per_rank == 1
     else:
         assert sampler.is_exact is False
-        assert sampler.n_chains == 16 * mpi.n_nodes * device_count_per_rank()
+        assert sampler.n_chains == 16 * mpi.n_nodes * jax.device_count()
 
 
 @common.skipif_distributed